image improvements
