+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-typescript",
   "version": "1.0.0",
-  "description": "React and TypeScript example starter project",
+  "description": "Tic Tac Toe Game in React and Typescript",
   "keywords": [
     "typescript",
     "react",
