+++ src/features/game/cell.scss
@@ -1,9 +1,29 @@
 .cell {
-    height: 40px;
-    flex-basis: 40px;
-    border-right: 1px solid #000;
-    border-bottom: 1px solid #000;
+    height: 12rem;
+    width: 12rem;
+    background-color: #b4cf8d;
+    border-radius: 1rem;
+    margin: 0.5rem;
     text-align: center;
+    font-size: 4rem;
+    box-shadow: 0 5px 5px;  // offset-x offset-y blur-radius spread-radius colors
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    
+    img {
+      width: 8rem;
+      height: 8rem;
+    }
+    img.mark-hover {
+      display: none
+    }
+    &:hover {
+      img.mark-hover {
+        display: flex;
+      }
+    }
   }
   
   .cell[data-id="22"] {
