+++ src/features/game/gameSettings.scss
@@ -14,8 +14,6 @@
     }
 }
 
-
-
 .btn-selection-container {
     background-color: #000;
     max-width: 20vw;
