+++ src/features/gameControls/gameControls.scss
@@ -1,0 +1,6 @@
+.btn-reset {
+    width: 4rem;
+    height: 4rem;
+    background-color: rgb(106, 187, 187);
+    border-radius: 1rem;
+}
